[REFACTOR] 마이 냉장고 폴더 분리 및 파일명 수정
