+++ src/components/MyRefrigerator/SelectedIngredient/SelectedList.tsx
